restore prod version for non-feature preview users
